Rename repo to opentelemetry-collector-components.
Fix import paths, and adopt the conventional directory structure for otel
components, nesting the oxide metrics receiver within receivers/.
