+++ receiver/oxidemetricsreceiver/config.go
@@ -1,4 +1,4 @@
-package oxidereceiver
+package oxidemetricsreceiver
 
 import (
 	"fmt"
