+++ receiver/oxidemetricsreceiver/config_test.go
@@ -1,4 +1,4 @@
-package oxidereceiver
+package oxidemetricsreceiver
 
 import (
 	"testing"
