+++ receiver/oxidemetricsreceiver/factory.go
@@ -1,4 +1,4 @@
-package oxidereceiver
+package oxidemetricsreceiver
 
 import (
 	"context"
